+++ db/migrate/20130222155658_add_csv_file_to_designs.rb
@@ -1,0 +1,5 @@
+class AddCsvFileToDesigns < ActiveRecord::Migration
+  def change
+    add_column :designs, :csv_file, :string
+  end
+end