+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.14.0
 
 ### Enhancements
+- **Design Changes**
+  - Designs along with data can now be imported from a CSV
 - **Gem Changes**
   - Updated to Contour 1.3.0.pre
   - Updated to Spreadsheet 0.8.1